refactor: improvded node builder methods
