+++ iroh/examples/rpc.rs
@@ -61,7 +61,7 @@ async fn run(db: impl Store) -> anyhow::Result<()> {
     // create a new node
     // we must configure the iroh collection parser so the node understands iroh collections
     let doc_store = iroh_sync::store::memory::Store::default();
-    let node = iroh::node::Node::builder(db, doc_store)
+    let node = iroh::node::Node::memory(db, doc_store)
         .secret_key(secret_key)
         .local_pool(&lp)
         .rpc_endpoint(rpc_endpoint)
