+++ iroh/src/node/builder.rs
@@ -1,4 +1,9 @@
-use std::{path::PathBuf, sync::Arc, time::Duration};
+use std::{
+    net::{Ipv4Addr, SocketAddrV4},
+    path::{Path, PathBuf},
+    sync::Arc,
+    time::Duration,
+};
 
 use anyhow::{bail, Context, Result};
 use futures::{FutureExt, StreamExt, TryFutureExt};
@@ -11,19 +16,24 @@ use iroh_bytes::{
 use iroh_gossip::net::{Gossip, GOSSIP_ALPN};
 use iroh_net::{derp::DerpMode, magic_endpoint::get_alpn, util::AbortingJoinHandle, MagicEndpoint};
 use iroh_sync::net::SYNC_ALPN;
-use quic_rpc::{transport::misc::DummyServerEndpoint, RpcServer, ServiceEndpoint};
+use quic_rpc::{
+    transport::{misc::DummyServerEndpoint, quinn::QuinnServerEndpoint},
+    RpcServer, ServiceEndpoint,
+};
 use serde::{Deserialize, Serialize};
 use tokio::sync::mpsc;
 use tokio_util::{sync::CancellationToken, task::LocalPoolHandle};
 use tracing::{debug, error, error_span, info, trace, warn, Instrument};
 
 use crate::{
+    client::quic::RPC_ALPN,
     node::{Event, NodeInner},
-    rpc_protocol::ProviderService,
+    rpc_protocol::{ProviderRequest, ProviderResponse, ProviderService},
     sync_engine::SyncEngine,
+    util::{fs::load_secret_key, path::IrohPaths},
 };
 
-use super::{rpc, Callbacks, DocStore, EventCallback, Node};
+use super::{rpc, Callbacks, DocStore, EventCallback, Node, RpcStatus};
 
 pub const PROTOCOLS: [&[u8]; 3] = [&iroh_bytes::protocol::ALPN, GOSSIP_ALPN, SYNC_ALPN];
 
@@ -60,7 +70,8 @@ where
     S: DocStore,
     E: ServiceEndpoint<ProviderService>,
 {
-    bind_port: u16,
+    root: Option<PathBuf>,
+    bind_port: Option<u16>,
     secret_key: SecretKey,
     rpc_endpoint: E,
     db: D,
@@ -73,11 +84,30 @@ where
     peers_data_path: Option<PathBuf>,
 }
 
+impl Default for Builder<iroh_bytes::store::mem::Store, iroh_sync::store::memory::Store> {
+    fn default() -> Self {
+        Self {
+            root: None,
+            bind_port: None,
+            secret_key: SecretKey::generate(),
+            db: Default::default(),
+            keylog: false,
+            derp_mode: DerpMode::Default,
+            rpc_endpoint: Default::default(),
+            gc_policy: GcPolicy::Disabled,
+            rt: None,
+            docs: Default::default(),
+            peers_data_path: None,
+        }
+    }
+}
+
 impl<D: Map, S: DocStore> Builder<D, S> {
     /// Creates a new builder for [`Node`] using the given database.
     pub fn with_db_and_store(db: D, docs: S) -> Self {
         Self {
-            bind_port: DEFAULT_BIND_PORT,
+            root: None,
+            bind_port: None,
             secret_key: SecretKey::generate(),
             db,
             keylog: false,
@@ -97,6 +127,41 @@ where
     S: DocStore,
     E: ServiceEndpoint<ProviderService>,
 {
+    /// Persist all node data in the provided directory.
+    pub async fn persist(
+        self,
+        root: impl AsRef<Path>,
+    ) -> Result<Builder<iroh_bytes::store::flat::Store, iroh_sync::store::fs::Store, E>> {
+        let root = root.as_ref();
+        let blob_dir = IrohPaths::BaoFlatStoreDir.with_root(root);
+        let peers_data_path = IrohPaths::PeerData.with_root(root);
+
+        tokio::fs::create_dir_all(&blob_dir).await?;
+        let root2 = root.to_path_buf();
+        tokio::task::spawn_blocking(|| migrate_flat_store_v0_v1(root2)).await??;
+        let bao_store = iroh_bytes::store::flat::Store::load(&blob_dir)
+            .await
+            .with_context(|| format!("Failed to load iroh database from {}", blob_dir.display()))?;
+        let doc_store = iroh_sync::store::fs::Store::new(IrohPaths::DocsDatabase.with_root(root))?;
+
+        let secret_key_path = IrohPaths::SecretKey.with_root(root);
+        let secret_key = load_secret_key(secret_key_path).await?;
+
+        Ok(Builder {
+            root: Some(root.into()),
+            bind_port: self.bind_port,
+            secret_key,
+            db: bao_store,
+            keylog: self.keylog,
+            rpc_endpoint: self.rpc_endpoint,
+            derp_mode: self.derp_mode,
+            gc_policy: self.gc_policy,
+            rt: self.rt,
+            docs: doc_store,
+            peers_data_path: Some(peers_data_path),
+        })
+    }
+
     /// Configure rpc endpoint, changing the type of the builder to the new endpoint type.
     pub fn rpc_endpoint<E2: ServiceEndpoint<ProviderService>>(
         self,
@@ -104,6 +169,7 @@ where
     ) -> Builder<D, S, E2> {
         // we can't use ..self here because the return type is different
         Builder {
+            root: self.root,
             bind_port: self.bind_port,
             secret_key: self.secret_key,
             db: self.db,
@@ -115,6 +181,31 @@ where
             docs: self.docs,
             peers_data_path: self.peers_data_path,
         }
+    }
+
+    /// Configure the default iroh rpc endpoint.
+    pub async fn enable_rpc(
+        self,
+    ) -> Result<Builder<D, S, QuinnServerEndpoint<ProviderRequest, ProviderResponse>>> {
+        let (ep, actual_rpc_port) = make_rpc_endpoint(&self.secret_key, DEFAULT_RPC_PORT)?;
+        if let Some(ref root) = self.root {
+            // store rpc endpoint
+            RpcStatus::store(root, actual_rpc_port).await?;
+        }
+
+        Ok(Builder {
+            root: self.root,
+            bind_port: self.bind_port,
+            secret_key: self.secret_key,
+            db: self.db,
+            keylog: self.keylog,
+            rpc_endpoint: ep,
+            derp_mode: self.derp_mode,
+            gc_policy: self.gc_policy,
+            rt: self.rt,
+            docs: self.docs,
+            peers_data_path: self.peers_data_path,
+        })
     }
 
     /// Sets the garbage collection policy.
@@ -143,7 +234,7 @@ where
     ///
     /// By default it binds to `127.0.0.1:11204`.
     pub fn bind_port(mut self, port: u16) -> Self {
-        self.bind_port = port;
+        self.bind_port.replace(port);
         self
     }
 
@@ -212,7 +303,8 @@ where
             Some(path) => endpoint.peers_data_path(path),
             None => endpoint,
         };
-        let endpoint = endpoint.bind(self.bind_port).await?;
+        let bind_port = self.bind_port.unwrap_or(DEFAULT_BIND_PORT);
+        let endpoint = endpoint.bind(bind_port).await?;
         trace!("created quinn endpoint");
 
         let (cb_sender, cb_receiver) = mpsc::channel(8);
@@ -512,6 +604,97 @@ async fn handle_connection<D: BaoStore>(
             .await
         }
         _ => bail!("ignoring connection: unsupported ALPN protocol"),
+    }
+    Ok(())
+}
+
+const DEFAULT_RPC_PORT: u16 = 0x1337;
+const MAX_RPC_CONNECTIONS: u32 = 16;
+const MAX_RPC_STREAMS: u32 = 1024;
+
+/// Makes a an RPC endpoint that uses a QUIC transport
+fn make_rpc_endpoint(
+    secret_key: &SecretKey,
+    rpc_port: u16,
+) -> Result<(QuinnServerEndpoint<ProviderRequest, ProviderResponse>, u16)> {
+    let rpc_addr = SocketAddrV4::new(Ipv4Addr::LOCALHOST, rpc_port);
+    let mut transport_config = quinn::TransportConfig::default();
+    transport_config
+        .max_concurrent_bidi_streams(MAX_RPC_STREAMS.into())
+        .max_concurrent_uni_streams(0u32.into());
+    let mut server_config = iroh_net::magic_endpoint::make_server_config(
+        secret_key,
+        vec![RPC_ALPN.to_vec()],
+        Some(transport_config),
+        false,
+    )?;
+    server_config.concurrent_connections(MAX_RPC_CONNECTIONS);
+
+    let rpc_quinn_endpoint = quinn::Endpoint::server(server_config.clone(), rpc_addr.into());
+    let rpc_quinn_endpoint = match rpc_quinn_endpoint {
+        Ok(ep) => ep,
+        Err(err) => {
+            if err.kind() == std::io::ErrorKind::AddrInUse {
+                tracing::warn!(
+                    "RPC port {} already in use, switching to random port",
+                    rpc_port
+                );
+                // Use a random port
+                quinn::Endpoint::server(
+                    server_config,
+                    SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0).into(),
+                )?
+            } else {
+                return Err(err.into());
+            }
+        }
+    };
+
+    let actual_rpc_port = rpc_quinn_endpoint.local_addr()?.port();
+    let rpc_endpoint =
+        QuinnServerEndpoint::<ProviderRequest, ProviderResponse>::new(rpc_quinn_endpoint)?;
+
+    Ok((rpc_endpoint, actual_rpc_port))
+}
+
+/// Migrate the flat store from v0 to v1. This can not be done in the store itself, since the
+/// constructor of the store now only takes a single directory.
+fn migrate_flat_store_v0_v1(iroh_data_root: PathBuf) -> anyhow::Result<()> {
+    let complete_v0 = iroh_data_root.join("blobs.v0");
+    let partial_v0 = iroh_data_root.join("blobs-partial.v0");
+    let meta_v0 = iroh_data_root.join("blobs-meta.v0");
+    let complete_v1 = IrohPaths::BaoFlatStoreDir
+        .with_root(&iroh_data_root)
+        .join("complete");
+    let partial_v1 = IrohPaths::BaoFlatStoreDir
+        .with_root(&iroh_data_root)
+        .join("partial");
+    let meta_v1 = IrohPaths::BaoFlatStoreDir
+        .with_root(&iroh_data_root)
+        .join("meta");
+    if complete_v0.exists() && !complete_v1.exists() {
+        tracing::info!(
+            "moving complete files from {} to {}",
+            complete_v0.display(),
+            complete_v1.display()
+        );
+        std::fs::rename(complete_v0, complete_v1).context("migrating complete store failed")?;
+    }
+    if partial_v0.exists() && !partial_v1.exists() {
+        tracing::info!(
+            "moving partial files from {} to {}",
+            partial_v0.display(),
+            partial_v1.display()
+        );
+        std::fs::rename(partial_v0, partial_v1).context("migrating partial store failed")?;
+    }
+    if meta_v0.exists() && !meta_v1.exists() {
+        tracing::info!(
+            "moving meta files from {} to {}",
+            meta_v0.display(),
+            meta_v1.display()
+        );
+        std::fs::rename(meta_v0, meta_v1).context("migrating meta store failed")?;
     }
     Ok(())
 }
